Day 15 Task Task 2
Solution for day 15 task 2, pathfinding on a larger grid. This is a very poor solution, as it is rather slow, and will not find the correct path if it requires backtracking into a previously cleared sector.

But it works, and I'm hungry, so I'm calling it a day and leaving this monstrosity as a reminder of my struggles and mistakes.