Day 15 Task 2
Implemented a system to find the empty space where we find all the small gaps(lines) between edges of exclusion zones.

Finding the intercept(s) of these lines gives us a point/points which are not covered by the exclusion zone of any of the sensors used to calculate the gaps.

As a final check, each gap is checked against the range of each sensor to ensure it is not within any exclusion zone.