2015 Day 13 Task 2
Updated the code from task 1 to calculate the happiness of each pairing. This is added up to get the total happiness, and the lowest scoring pair is tracked to find the best place to insert "myself" as this nullifies the given pairing.

For this input, the best layout with me is found by inserting myself into the existing best layout, but the code should work for instances where this is not the case.