2024 Day 5 Task 2
Not sure of the time complexity for sorting the invalid lists.

All lists for the input only ran the main loop of "FixOrder()" once (And therefore never added any late values back into "pagesToAdd" list).

It is possible to have inputs that do make use of that section, and therefore loop the main function more times. This therefore increases the execution time.