2024 Day 9 Task 2 (Slow)
Task 2 section is slow because it needs to iterate through many elements of the map for each data chunk.

This was the used solution as initially it was implemented to just check the first entry for each key to find the earliest possible gap.

This solution did not work, as maps are only sorted by key, not by value. So when inserting gaps into the map *after* the initial read-in, these new gaps were added to the end, often out of order.

Optimisation would be to use an array of 9 sorted containers to track the gaps, as this would work as originally intended with only checking the first entry for each appropriate gap.